src/app/shared/components/dropdown/dropdown.component.ts
import { Component, ElementRef, HostListener } from '@angular/core';

@Component({
  selector: 'app-dropdown',
  templateUrl: './dropdown.component.html',
  styleUrl: './dropdown.component.scss',
})
export class DropdownComponent {
  protected isOpen = false;

  constructor(private elementRef: ElementRef) {}

  toggleDropdown() {
    this.isOpen = !this.isOpen;
  }

  @HostListener('document:click', ['$event.target'])
  onClickOutside(target: HTMLElement): void {
    if (!this.isOpen) return;

    const dropdown = this.elementRef.nativeElement;

    if (!dropdown?.contains(target)) {
      this.isOpen = false;
    }
  }
}
